requirements.txt
boto3==1.34.34
python-dotenv==1.0.1
tqdm==4.66.1
loguru==0.7.2
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
motor==3.3.2
pymongo==4.6.1 